+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.3.4 - 2024-10-21
+* [#15](https://github.com/alipay/global-open-sdk-php/pull/15) feature-241022
+  - add AlipayVaultingSessionRequest.is3DSAuthentication哈啰集成绑卡SDK
+  - add AlipayAuthConsultRequest.authMetaData AE企业支付宝代扣集成方案
+  - add Boolean  request.order.needDeclaration
 
 ## 1.3.3 - 2024-09-14
 * [#14](https://github.com/alipay/global-open-sdk-php/pull/14)  update pay